+++ CHANGELOG.rst
@@ -7,6 +7,7 @@ Changelog
 https://gitlab.com/gf-bl/oereb_client/milestones/16
 
 - Adjust legend titles according to PDF specification
+- Add responsible offices from geometries
 
 1.2.3
 *****